pyproject.toml
[tool.poetry]
name = "learn-langgraph"
version = "0.1.0"
description = ""
authors = ["alan <yangxuangui1117@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langgraph = "^0.2.70"
langchain-deepseek = "^0.1.1"
langchain-core = "^0.3.34"
python-dotenv = "^1.0.1"
pytest = "^8.3.4"
langchain-community = "^0.3.17"
langchain-anthropic = "^0.3.7"
tavily-python = "^0.5.1"
pandas = "^2.2.3"
openai = "^1.61.1"
langsmith = "^0.3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
